+++ package.json
@@ -50,12 +50,12 @@
     "mocha": "^6.2.1",
     "power-assert": "^1.6.0",
     "rimraf": "^3.0.0",
-    "typescript": "~3.6.0"
+    "typescript": "~3.6.0",
+    "@types/fs-extra": "^8.0.1",
+    "fs-extra": "^8.1.0"
   },
   "dependencies": {
-    "@types/fs-extra": "^8.0.1",
     "command-line-args": "^5.0.2",
-    "fs-extra": "^8.1.0",
     "get-stdin": "^7.0.0",
     "nunjucks": "^3.1.3",
     "protobufjs": "^6.8.8"
